Added new feature

+++ main.py
@@ -2,23 +2,26 @@ import os, sys
 args = sys.argv[1:]
 
 def main():
-    code = []
-    with open(args[0]) as f:
-        for line in f:
-            code.append(line.strip("\n"))
-        filename = args[0].split("/")[-1].split(".")[0]
+    if os.path.exists(args[0]):
+        code = []
+        with open(args[0]) as f:
+            for line in f:
+                code.append(line.strip("\n"))
+            filename = args[0].split("/")[-1].split(".")[0]
 
-    with open(f"{filename}.tmpbash", 'w') as f:
-        f.write("#!/bin/bash\n\npython3 -c \"\n")
-        for line in code:
-            line = line.replace("\\", "\\\\").replace("\"", "\\\"")
-            f.write(f"{line}\n")
-        f.write("\n\" $@")
+        with open(f"{filename}.tmpbash", 'w') as f:
+            f.write("#!/bin/bash\n\npython3 -c \"\n")
+            for line in code:
+                line = line.replace("\\", "\\\\").replace("\"", "\\\"")
+                f.write(f"{line}\n")
+            f.write("\n\" $@")
 
-    os.system(f"shc -f {filename}.tmpbash")
-    os.remove(f"{filename}.tmpbash")
-    os.remove(f"{filename}.tmpbash.x.c")
-    os.rename(f"{filename}.tmpbash.x", f"{filename}.binary")
+        os.system(f"shc -f {filename}.tmpbash")
+        os.remove(f"{filename}.tmpbash")
+        os.remove(f"{filename}.tmpbash.x.c")
+        os.rename(f"{filename}.tmpbash.x", f"{filename}.binary")
+    else:
+        print("ERROR: File not found. ")
 
 
 def help():
